refactored process current record to tech record repository
