+++ src/services/TechRecordsRepository.ts
@@ -4,7 +4,7 @@ import { InvocationRequest } from '@aws-sdk/client-lambda';
 import { IInvokeConfig } from '../models/IInvokeConfig';
 import { Configuration } from '../utils/Configuration';
 import { LambdaService } from './LambdaService';
-import { TechRecordGet, TechRecordType } from '../models/Types';
+import { ISearchResult, TechRecordGet, TechRecordType } from '../models/Types';
 
 @Service()
 export class TechRecordsRepository {
@@ -76,5 +76,55 @@ export class TechRecordsRepository {
       console.log(JSON.stringify(e));
       return undefined;
     }
+  };
+
+  public processGetCurrentProvisionalRecords = async <T extends TechRecordGet['techRecord_vehicleType']>(searchResult: ISearchResult[]): Promise<TechRecordType<T> | undefined> => {
+    if (searchResult) {
+      const processRecordsRes = this.groupRecordsByStatusCode(searchResult);
+
+      if (processRecordsRes.currentCount !== 0) {
+        return this.callGetTechRecords(
+          processRecordsRes.currentRecords[0].systemNumber,
+          processRecordsRes.currentRecords[0].createdTimestamp,
+        );
+      }
+
+      if (processRecordsRes.provisionalCount === 1) {
+        return this.callGetTechRecords(
+          processRecordsRes.provisionalRecords[0].systemNumber,
+          processRecordsRes.provisionalRecords[0].createdTimestamp,
+        );
+      }
+
+      return this.callGetTechRecords(
+        processRecordsRes.provisionalRecords[1].systemNumber,
+        processRecordsRes.provisionalRecords[1].createdTimestamp,
+      );
+    }
+
+    return Promise.reject(new Error('Tech record Search returned nothing.'));
+  };
+
+  /**
+   * helper function is used to process records and count provisional and current records
+   * @param records
+   */
+  private groupRecordsByStatusCode = (records: ISearchResult[]): { currentRecords: ISearchResult[]; provisionalRecords: ISearchResult[]; currentCount: number; provisionalCount: number; } => {
+    const currentRecords: ISearchResult[] = [];
+    const provisionalRecords: ISearchResult[] = [];
+    records.forEach((record) => {
+      if (record.techRecord_statusCode === 'current') {
+        currentRecords.push(record);
+      } else if (record.techRecord_statusCode === 'provisional') {
+        provisionalRecords.push(record);
+      }
+    });
+
+    return {
+      currentRecords,
+      provisionalRecords,
+      currentCount: currentRecords.length,
+      provisionalCount: provisionalRecords.length,
+    };
   };
 }
